pyproject.toml
[project]
name = "osprey-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.77.1",
    "atkafka-consumer-python",
    "atproto>=0.0.65",
    "click>=8.3.1",
    "clickhouse-connect>=0.10.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]

[tool.uv.sources]
atkafka-consumer-python = { git = "https://tangled.org/hailey.at/atkafka-consumer-python" }
